refactor: move PhysicalTenantContext in common gateway module
PhysicalTenantInterceptor now use PhysicalTenantResolver
instead of PhysicalTenantRegistry
