+++ gateways/gateway-mapping-http/src/test/java/io/camunda/gateway/mapping/http/physicaltenants/PhysicalTenantContextTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.rest.context;
+package io.camunda.gateway.mapping.http.physicaltenants;
 
 import static org.assertj.core.api.Assertions.assertThat;
